[CustomOp] restructure class methods from class hierachy
