+++ src/finn/custom_op/fpgadataflow/rtl/thresholding_rtl.py
@@ -154,128 +154,6 @@ class Thresholding_rtl(Thresholding, RTLBackend):
         res_dict = self.get_memory_estimate()
         return res_dict.get("LUTRAM", 0)
 
-    def get_input_datatype(self, ind=0):
-        return DataType[self.get_nodeattr("inputDataType")]
-
-    def get_output_datatype(self, ind=0):
-        return DataType[self.get_nodeattr("outputDataType")]
-
-    def get_weight_datatype(self):
-        """The term 'weights' and 'thresholds' are used interchangably in this class."""
-        return DataType[self.get_nodeattr("weightDataType")]
-
-    def minimize_accumulator_width(self, model):
-        "Minimize threshold width ('accumulator width' here due to convention)"
-        thresholds = model.get_initializer(self.onnx_node.input[1])
-        threshold_tensor = self.get_hw_compatible_threshold_tensor(thresholds)
-        min_threshold = thresholds.min()
-        max_threshold = thresholds.max()
-        min_input = self.get_input_datatype().min()
-        max_input = self.get_input_datatype().max()
-        # get range required by threshold values
-        tdt_min = min(min_input, min_threshold)
-        tdt_max = max(max_input, max_threshold)
-        if tdt_min < 0:
-            if abs(tdt_min) > tdt_max:
-                tdt = DataType.get_smallest_possible(tdt_min)
-            else:
-                tdt = DataType.get_smallest_possible(-tdt_max - 1)
-        else:
-            tdt = DataType.get_smallest_possible(tdt_max)
-        assert np.vectorize(tdt.allowed)(
-            threshold_tensor
-        ).all(), "Thresholds can't be expressed with type %s" % str(tdt)
-        self.set_nodeattr("weightDataType", tdt.name)
-        return DataType[self.get_nodeattr("weightDataType")]
-
-    def get_instream_width(self, ind=0):
-        i_bits = self.get_input_datatype().bitwidth()
-        return i_bits * self.get_nodeattr("PE")
-
-    def get_outstream_width(self, ind=0):
-        o_bits = self.get_output_datatype().bitwidth()
-        return o_bits * self.get_nodeattr("PE")
-
-    def get_weightstream_width(self):
-        """Returns weight stream width"""
-        pe = self.get_nodeattr("PE")
-        wp = self.get_weight_datatype().bitwidth()
-        n_thres_steps = self.get_nodeattr("numSteps")
-        w_width = pe * wp * n_thres_steps
-        return w_width
-
-    def get_folded_input_shape(self, ind=0):
-        fold = self.calc_tmem()
-        pe = self.get_nodeattr("PE")
-        vecs = list(self.get_nodeattr("numInputVectors"))
-        folded_input_shape = tuple(vecs + [fold, pe])
-        return folded_input_shape
-
-    def get_folded_output_shape(self, ind=0):
-        # same shape as input
-        return self.get_folded_input_shape()
-
-    def get_normal_input_shape(self, ind=0):
-        num_channels = self.get_nodeattr("NumChannels")
-        vecs = list(self.get_nodeattr("numInputVectors"))
-        normal_input_shape = tuple(vecs + [num_channels])
-        return normal_input_shape
-
-    def get_normal_output_shape(self, ind=0):
-        # same shape as input
-        return self.get_normal_input_shape()
-
-    def get_number_output_values(self):
-        return np.prod(self.get_folded_output_shape()[:-1])
-
-    def get_exp_cycles(self):
-        # Channels/PE * batch size * fmdim * fmdim
-        return np.prod(self.get_folded_output_shape()[:-1])
-
-    def get_hw_compatible_threshold_tensor(self, orig_thres_matrix):
-        """Convert the original numpy weight matrix orig_weight_matrix into
-        a form suitable for passing to the hlslib call:
-        * ensure MH % PE == 0
-        * for unsigned inputs, ensure thresholds are positive
-        * interleave rows between PEs
-        * reshape into (PE, TMEM, n_thres_steps) and return
-        """
-        mh = self.get_nodeattr("NumChannels")
-        pe = self.get_nodeattr("PE")
-        tmem = mh // pe
-        assert mh % pe == 0, "Requirement NumChannels divisable by PE is violated."
-        assert (
-            orig_thres_matrix.ndim == 2
-        ), """Threshold matrix dimension is
-        not as expected (2)."""
-        n_thres_steps = orig_thres_matrix.shape[1]
-        assert n_thres_steps == self.get_nodeattr("numSteps"), "Mismatch in threshold steps"
-        if not self.get_input_datatype().signed():
-            # ensure all thresholds are nonnegative
-            assert (orig_thres_matrix >= 0).all()
-        # ensure all thresholds are integer
-        assert np.equal(np.mod(orig_thres_matrix, 1), 0).all(), "Need int threshold tensor"
-        ret = orig_thres_matrix
-        # ensure channels = mh , duplicating if necessary
-        if ret.shape[0] == 1:
-            ret = np.tile(ret, (mh, 1))
-        assert ret.shape[0] == mh, "Channels of threshold matrix are not as expected (mh)"
-        # distribute rows between PEs
-        ret = interleave_matrix_outer_dim_from_partitions(ret, pe)
-        assert (
-            ret.shape[0] == pe
-        ), """First dimension after distribution of the
-        rows between PEs is not as expected (pe)"""
-        assert (
-            ret.shape[1] == tmem
-        ), """Second dimension after distribution of the
-        rows between PEs is not as expected (tmem)"""
-        assert (
-            ret.shape[2] == n_thres_steps
-        ), """Third dimension after distribution of the
-        rows between PEs is not as expected (n_thres_steps)"""
-        return ret.reshape(1, pe, tmem, n_thres_steps)
-
     def get_all_meminit_filenames(self, abspath=False):
         "Return a list of all .dat memory initializer files used for this node"
         dat_files = []
@@ -623,23 +501,7 @@ class Thresholding_rtl(Thresholding, RTLBackend):
         return cmd
 
     def get_verilog_top_module_intf_names(self):
-        """Return a dict of names of input and output interfaces.
-        The keys reflect the protocols each interface implements:
-        'clk', 'rst', 'm_axis', 's_axis', 'aximm', 'axilite'.
-        Values are lists of tuples (axis, aximm) or names (axilite):
-        'axis' tuples correspond to the list of node inputs in order,
-        each tuple is (interface_name, interface_width_bits).
-        axilite always assumed to be 32 bits and is not tuple (name only).
-        Each block must have at most one aximm and one axilite."""
-
-        intf_names = {}
-        intf_names["clk"] = ["ap_clk"]
-        intf_names["rst"] = ["ap_rst_n"]
-        intf_names["s_axis"] = [("in0_V", self.get_instream_width_padded())]
-        intf_names["m_axis"] = [("out_V", self.get_outstream_width_padded())]
-        intf_names["aximm"] = []
-        intf_names["axilite"] = []
-        intf_names["ap_none"] = []
+        intf_names = super().get_verilog_top_module_intf_names()
         if self.get_nodeattr("runtime_writeable_weights") == 1:
             intf_names["axilite"] = ["s_axilite"]
